Add option to list processes without killing them

+++ kill-on-port.fish
@@ -3,6 +3,7 @@
 #
 # USAGE
 #   Options
+#       -l, --list     List processes using the specified ports (does not kill)
 #       -r, --root     Attempts to kill the process as root
 #       -f, --force    Does not ask for confirmation
 #       -h, --help     Prints usage
@@ -13,14 +14,17 @@ end
 
 function kill-on-port -d "Kills a process that is using the given port"
   set -l force 0
+  set -l list 0
   set -l root ""
 
-  while set optarg (getopts "f:force h:help r:root" $argv)
+  while set optarg (getopts "f:force h:help r:root l:list" $argv)
     switch $optarg[1]
       case f
         set force 1
       case r
         set root "sudo "
+      case l
+        set list 1
       case h
         __kill-on-port.print-usage
         return
@@ -41,20 +45,22 @@ function kill-on-port -d "Kills a process that is using the given port"
         set -l name (echo $process | awk '{print $1}')
         set -l pid (echo $process | awk '{print $2}')
         echo "Found process named $name with process id $pid using port $port"
-        if test $force = 1
-          eval "$root kill -9 $pid"
-          echo "Successfully killed $pid"
-        else
-          echo "Continue (y/n):"
-          read -l choice
-          switch $choice
-            case y yes Y YES
-              eval "$root kill -9 $pid"
-              echo "Successfully killed $pid"
-            case n no N NO
-              echo "Not killing $pid"
-            case '*'
-              echo "Unrecognized choice: $choice"
+        if test $list = 0
+          if test $force = 1
+            eval "$root kill -9 $pid"
+            echo "Successfully killed $pid"
+          else
+            echo "Continue (y/n):"
+            read -l choice
+            switch $choice
+              case y yes Y YES
+                eval "$root kill -9 $pid"
+                echo "Successfully killed $pid"
+              case n no N NO
+                echo "Not killing $pid"
+              case '*'
+                echo "Unrecognized choice: $choice"
+            end
           end
         end
